workaround windows 8192 stdin with binary tempfile

+++ src/cglow.f90
@@ -46,7 +46,7 @@ module cglow
 !    >                THI(NEI,NMAJ),  AK(NEI,NMAJ),   AJ(NEI,NMAJ),
 !    >                TS(NEI,NMAJ),   TA(NEI,NMAJ),   TB(NEI,NMAJ),
 !    >                GAMS(NEI,NMAJ), GAMB(NEI,NMAJ)
-
+  use, intrinsic :: iso_fortran_env, only: wp=>real32
 
   implicit none
   save
@@ -78,7 +78,7 @@ module cglow
   real,allocatable,dimension(:) ::             &                   ! (jmax)
     zz, zo, zn2, zo2, zno, zns, znd, zrho, ze, &
     ztn, zti, zte, eheat, tez, ecalc, tei, tpi, tir
-  real,allocatable,dimension(:)     :: phitop, ener, del           ! (nbins)
+  real(wp),allocatable,dimension(:)     :: phitop, ener, del           ! (nbins)
   real,allocatable,dimension(:)     :: wave1, wave2, sflux         ! (lmax)
   real,allocatable,dimension(:,:)   :: pespec, sespec, uflx, dflx  ! (nbins,jmax)
   real,allocatable,dimension(:,:)   :: zmaj, zcol, pia, sion       ! (nmaj,jmax)
